Componentize, update gitignore
